+++ vignettes/MeasureRadii/aaaMakeData4Vignettes.R
@@ -4,12 +4,15 @@ setwd("C:/aaaWork/Programs/GitHub/RFishBC/vignettes/MeasureRadii")
 
 ## Process Scales
 junk <- digitizeRadii("Scale_1.jpg",id="1",reading="DHO",edgeIsAnnulus=FALSE)
-digitizeRadii("Scale_1.jpg",id="1",reading="DHO2",edgeIsAnnulus=FALSE)
+## choose Scale_1.jpg and set id to 1
+digitizeRadii(reading="DHO2",edgeIsAnnulus=FALSE,popID=TRUE)
 digitizeRadii("Scale_1.jpg",id="1",reading="DHO3",edgeIsAnnulus=FALSE)
 digitizeRadii("Scale_2.jpg",id="2",reading="DHO",edgeIsAnnulus=FALSE)
 
 showDigitizedImage("Scale_1_DHO.RData")
 showDigitizedImage(c("Scale_1_DHO.RData","Scale_1_DHO2.RData","Scale_1_DHO3.RData"))
+
+
 
 ## Process the otolith
 digitizeRadii("Oto140306.jpg",id="140306",reading="DHO",